+++ src/main/java/com/growup/pms/auth/service/oauth/OAuth2Service.java
@@ -1,4 +1,4 @@
-package com.growup.pms.auth.service;
+package com.growup.pms.auth.service.oauth;
 
 import com.growup.pms.auth.service.dto.oauth.OAuthAccessToken;
 import com.growup.pms.auth.service.dto.oauth.OAuthProfile;
