+++ src/main/java/com/growup/pms/auth/service/oauth/OAuth2ServiceImpl.java
@@ -1,6 +1,7 @@
-package com.growup.pms.auth.service;
+package com.growup.pms.auth.service.oauth;
 
-import static org.springframework.http.HttpHeaders.*;
+import static org.springframework.http.HttpHeaders.AUTHORIZATION;
+import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
