+++ src/main/java/com/growup/pms/auth/service/oauth/OAuthLoginService.java
@@ -1,6 +1,7 @@
-package com.growup.pms.auth.service;
+package com.growup.pms.auth.service.oauth;
 
 import com.growup.pms.auth.controller.dto.SecurityUser;
+import com.growup.pms.auth.service.RefreshTokenService;
 import com.growup.pms.auth.service.dto.oauth.OAuthAccessToken;
 import com.growup.pms.auth.service.dto.oauth.OAuthProfile;
 import com.growup.pms.common.exception.code.ErrorCode;
